APP-71: [PDex]: Update logic show warning.
