+++ src/screens/DexV2/components/Trade/utils.js
@@ -1,5 +1,8 @@
 import _ from 'lodash';
 import { COINS } from '@src/constants';
+import { useSelector } from 'react-redux';
+import { selectedPrivacySeleclor } from '@src/redux/selectors';
+import BigNumber from 'bignumber.js';
 
 export const calculateOutputValueCrossPool = (pairs, inputToken, inputValue, outputToken) => {
   const firstPair = _.get(pairs, 0);
@@ -48,4 +51,48 @@ export const calculateInputValue = (pair, inputToken, outputValue, outputToken) 
   } catch (error) {
     console.debug('CALCULATE OUTPUT', error);
   }
+};
+
+const convertToUsdNumber = (multiple, multipliedBy, decimal) => {
+  return BigNumber(multiple)
+    .multipliedBy(BigNumber(multipliedBy))
+    .dividedBy(BigNumber(10).pow(decimal))
+    .toNumber() || 0;
+};
+
+const getImpact = (input, output) => {
+  input   = BigNumber(input);
+  output  = BigNumber(output);
+  return output
+    .minus(input)
+    .dividedBy(input)
+    .decimalPlaces(3)
+    .toNumber();
+};
+
+export const calculateSizeImpact = (inputValue, inputToken, outputValue, outputToken) => {
+  const {
+    priceUsd:   inputPriceUsd,
+    pDecimals:  inputPDecimals
+  } = useSelector(selectedPrivacySeleclor.getPrivacyDataByTokenID)(inputToken?.id);
+  const {
+    priceUsd:   outputPriceUsd,
+    pDecimals:  outputPDecimals
+  } = useSelector(selectedPrivacySeleclor.getPrivacyDataByTokenID)(outputToken?.id);
+  const totalInputUsd   = convertToUsdNumber(inputValue, inputPriceUsd, inputPDecimals);
+  const totalOutputUsd  = convertToUsdNumber(outputValue, outputPriceUsd, outputPDecimals);
+  if (totalInputUsd && totalInputUsd !== 0) {
+    const impact = getImpact(totalInputUsd, totalOutputUsd);
+    if (!isNaN(impact)) {
+      return {
+        impact: impact,
+        showWarning: impact < -5
+      };
+    }
+  }
+
+  return {
+    impact: null,
+    showWarning: false
+  };
 };
